backend/pyproject.toml
[tool.poetry]
name = "changelog"
version = "0.1.0"
description = ""
authors = ["Marcin Włodarczyk <gonzur75@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
pre-commit = "^3.8.0"
fastapi = {extras = ["standard"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
pytest = "^8.3.2"
pymysql = "^1.1.1"
cryptography = "^43.0.0"
coverage = "^7.6.1"
bcrypt = "^4.2.0"
faker = "^28.0.0"
pyjwt = "^2.9.0"
factory-boy = "^3.3.1"
mypy = "^1.11.2"
alembic = "^1.13.2"
ruff = "^0.6.3"
tenacity = "^9.0.0"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core==1.8.3"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]
